pyproject.toml
[project]
name = "repl-nix-cardcollectorhub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.39.0",
]
